+++ CHANGELOG.md
@@ -267,3 +267,4 @@
 * Remove `Schematic` module and move `Schematic` type family to `Interpreted` module
 * Change arity of `Schematic` type family
 * Remove `lay` method of `Monadic` type class, use `lift` instead
+* Remove `flick` method of `Comonadic` type class, use `lower` instead